Move risk module into (risks) route group
